fix(get-stage-changes): review fixes - move use case `GetStageChanges` to profile `recruitment/get-stage-changes`
