+++ grid/recruitment/candidates/maps/breezy-hr.suma
@@ -28,13 +28,6 @@ map CreateCandidate {
   }
 }
 
-// TODO: implement this
-map GetStageChanges {
-  map result {
-    stages: []
-  }
-}
-
 // Breezy HR rate limits docs: https://developer.breezy.hr/reference/rate-limiting
 operation MapRateLimit {
   totalRequests = parseInt(args.headers['x-ratelimit-limit'], 10)
