+++ grid/recruitment/candidates/maps/mock.suma
@@ -10,27 +10,3 @@ map CreateCandidate {
     jobId = 'JOB_ID'
   }
 }
-
-"""
-GetStageChanges map
-"""
-map GetStageChanges {
-  return map result {
-    stages = [
-      {
-        id: 'sourced',
-        name: 'Sourced',
-        description: 'Sourced from linkedIn',
-        current: false,
-        createdAt: "2022-12-06T15:20:11Z",
-      },
-      {
-        id: 'applied',
-        name: 'Applied',
-        description: 'Applied for job opening of Software Engineer',
-        current: true,
-        createdAt: "2022-12-08T10:20:42Z",
-      }
-    ]
-  }
-}
