+++ grid/recruitment/candidates/profile.supr
@@ -141,108 +141,6 @@ usecase CreateCandidate {
 }
 
 """
-Get Stage Changes
-Retrieves a list of stages for specific candidate throughout his recruitment process. The candidate does not have to be applied for a job.
-"""
-usecase GetStageChanges {
-  input {
-    """
-    Candidate Id
-    Unique identifier of a candidate.
-    """
-    candidateId! string!
-  }
-
-  result {
-    """
-    Stages
-    List of stages relevant to the candidate.
-    """
-    stages! [Stage!]!
-    
-    """
-    Rate limit
-    Rate limit details.
-    """
-    rateLimit RateLimit
-  }!
-
-  error Error
-
-  example Successful {
-    input {
-      candidateId = 'CANDIDATE_ID'
-    }
-
-    result {
-      stages = [
-        {
-          id = 'sourced',
-          name = 'Sourced',
-          description = 'Sourced from LinkedIn',
-          current = false,
-          createdAt = "2022-12-06T15:20:11Z",
-        },
-        {
-          id = 'applied',
-          name = 'Applied',
-          description = 'Applied for job opening of Software Engineer',
-          current = true,
-          createdAt = "2022-12-08T10:20:42Z",
-        }
-      ]
-    }
-  }
-
-  example Failed {
-    input {
-      candidateId = 'NOT_EXISTING_CANDIDATE_ID'
-    }
-
-    error {
-      title = 'Not found'
-      detail = 'Not found'
-      code = 'NotFound'
-      rateLimit = {
-        bucket = 'workable'
-      }
-    }
-  }
-}
-
-model Stage {
-  """
-  Id
-  Unique identifier of a stage
-  """
-  id! string!
-
-  """
-  Name
-  Human readable name of a stage.
-  """
-  name! string!
-
-  """
-  Description
-  Description of a stage.
-  """
-  description string
-
-  """
-  Current
-  Whether candidate is in this stage currently.
-  """
-  current boolean
-
-  """
-  Created At
-  ISO8601 timestamp when candidate moved to this stage.
-  """
-  createdAt! string!
-}
-
-"""
 Education
 Object representing educational experience of a candidate.
 """
