+++ superface/super.json
@@ -816,6 +816,14 @@
           "file": "../grid/recruitment/jobs/maps/breezy-hr.suma"
         }
       }
+    },
+    "recruitment/get-stage-changes": {
+      "file": "../grid/recruitment/get-stage-changes/profile.supr",
+      "providers": {
+        "mock": {
+          "file": "../grid/recruitment/get-stage-changes/maps/mock.suma"
+        }
+      }
     }
   },
   "providers": {
